Move collapsed-borders conflict resolution logic to layout.table module
